fix: DB 칼럼 타입 수정

+++ src/main/java/com/osondoson/backend/domain/player/dto/SeasonHistory.java
@@ -12,8 +12,8 @@ public record SeasonHistory(
         String club,
         String league,
         String clubLogoUrl,
-        Double appearances,
-        Double minutes,
+        Integer appearances,
+        Integer minutes,
         Double ratingAverage,
         List<KeyStat> keyStats
 ) {
